+++ app/Models/Inbox.php
@@ -2,7 +2,9 @@
 
 namespace App\Models;
 
+use App\Models\Traits\HasActivityLogs;
 use Illuminate\Database\Eloquent\Casts\Attribute;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 use Illuminate\Database\Eloquent\SoftDeletes;
@@ -10,7 +12,7 @@ use Illuminate\Support\Facades\Auth;
 
 class Inbox extends Model
 {
-    use SoftDeletes;
+    use HasActivityLogs, HasFactory, SoftDeletes;
 
     protected $table = 'fm_inboxes';