+++ database/migrations/2025_03_07_141559_create_fm_messages_table.php
@@ -1,9 +1,9 @@
 <?php
 
+use App\Models\Inbox;
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
-use Jeddsaliba\FilamentMessenger\Models\Inbox;
 
 return new class extends Migration
 {